linux/linux-mainline-rt: refactor do_firmware_path
do_firmware_path makes sure that the path set in CONFIG_EXTRA_FIRMWARE
is correct changing the path with sed.

Make this path modification done on .config rather than defconfig so
that the root configuration is never altered.

Execute this task later after the configuration is done and just before
building so that we avoid any conflicts with other Yocto Kernel stacks.

Signed-off-by: Eloi Bail <eloi.bail@savoirfairelinux.com>