wic/mkefidisk-guest.wks.in generate guest image in qcow2 format
Use the wic.qcow2 in order to generate the guest image. wic.qcow2 use
wic and preserves wic partitions and kernel parameters but generates a
qcow2 image directly usable by qemu.

Note: the generated image is an UEFI bootable image. So libvirt must be
configured in UEFI in order to use this image.

To configure libvirt for use with UEFI add an attribute firmware="efi"
in os xml tag.

You can use this example:
<os firmware="efi">
    <type arch="x86_64" machine="pc-i440fx-4.1">hvm</type>
    <boot dev="hd"/>
    <bootmenu enable="no"/>
    <bios useserial="yes" rebootTimeout="0"/>
    <smbios mode="emulate"/>
</os>

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>