rte-host-common.inc: install a complete cluster solution
Install crmsh, openssh and cluster-glue.

crmsh is a tool to configure a HA cluster based and corosync and
pacemaker. We used it facilitate pacemaker configuration.

Also add openssh, crmsh uses ssh to distribute configurations across
nodes.

Cluster Glue is a set of libraries, tools and utilities suitable for
the Heartbeat/Pacemaker cluster stack. We need it for STONITH (fencing)
support.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>