recipes-votp/system-config: install public ssh key
Add a public key to support root keys ssh connection for Ansible.
Ansible needs a ssh access on the machine in order to work.

The public key must be placed in
yocto-bsp/keys/ansible_public_ssh_key.pub. A symlink in system-config
recpies allows to get the key outside of the Yocto build system.
If there is no ssh public key available, the build will failed.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>