conf: classes/security: create hardening framework
Initiate the hardening framework that will be the environment in which
all Cyber-security related modifications will be included and
controlled.

This framework is intended to evolve towards a generic mechanism that
assists developer in implementing checks, audits, quality gates and
data collection / document generation that will be used for
Cyber-security purposes.

SFL: #5514
Change-Id: I03ac27118c4666b7efedf9c2ae3bf5edede4253a
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>