classes/security: add manifests-archiver class
Add a bbclass that archives any artifact produced while building
an image and that could be required as a proof of implementation
or analysis by security audit representatives.

This commit sets up the environment and simply packages the list
of installed apps and CVEs in a tarball.

It is also intended to serve as a template for upcoming security
classes that will be implemented.

Issue: #5514
Change-Id: I5f9a1c772c9e9e3e48ef840576da6a49ab9a76dd
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>