openvswitch: launch service if iommu is enabled
Add a wConditionDirectoryNotEmpty which checks that iommu is enabled by
monitoring the /sys/class/iommu directory, and starts the
openvswitch.service if it is.
This is required as using openvswitch without iommu make dpdk load the
`uio_pci_generic` driver, which uses physical device adresses and
require broader filesystem accesses than the `vfio-pci` driver.
Thus, to increase security and reduce the access of openvswitch and dpdk
to the filesystem, force the use of `vfio-pci`.

Issue: #5335
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Change-Id: Ib43762a6ac5aa65e6ea55a393fc363f65eca522d