cukinia-tests: fix ceph tests to prevent false positives
Some tests were checking against the number of lines returned by
"ceph status" function, which is weak against a failure of the
aforementioned command.

Add a wrapper function to run the status once and detect if the
command has failed or not. Returns the output if the command
succeeded or an error otherwise.

Issue: #5334
Change-Id: Id0795fcf84df50d1dd5eef753e3cf94c3eb1e764
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>