wic: cukinia-tests: disable audit subsystem
Audit subsystem is commonly used by "auditd" or SELinux to alert on a
collection of events triggered during the system's lifecycle.

Disable that subsystem on the cmdline as it is unused for by the host
system.

Do not disable audit subsystem for guest images for now as there is
no status about the usage of that subsystem in that kind of image.

Issue: #5338
Change-Id: I231cc6ae427c5f3be9a64636f28991eb8562bf5c
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>