classes/security: add rootfs-tweaks
Add rootfs-tweaks.bbclass, a class used to harden the rootfs by
modifying it after its generation.
Add a random-root-passwd service used to randomize the root password at
each boot. This service is not installed when the debug-tweaks image
feature is used, as debug-tweaks sets an empty root password for
debugging purpose.

Issue: #5337
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Change-Id: Ie72474fd476eed444f89b8f0d2b8b6b4703c909e