security: add support for installing hardened PAM policy
PAM configuration is a pretty orthogonal piece of configuration which
has a lot of impact on application's behavior while authenticating.

To improve auditability of the PAM configuration, set it all in a
dedicated folder under the security framework and install each
configuration file before preparing the image. This policy will
overwrite any defined configuration in the rootfs.

Also migrate already existing PAM configurations to this new
artifact space.

Issue: #5803
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>
Change-Id: I3311177b3be6f1f62ff44a4c16e54dc9240c2984