security: enforce a unique keyring on session creation
pam_keyinit is a module that assists in instantiating a Kernel session
keyring for the application that is attempting to log in.

Sharing a keyring between two authenticated applications introduce
possibility of having one application stealing credentials from
the other.

Having it as "requisite" ensure that this is a mandatory condition for
authentication : if the application cannot have a distinct keyring,
the authentication must be denied.

Issue: #5803
Change-Id: I089ebb52d11a43ea70cc79655eff2bf5c1ce51c4
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>