classes: images: security: install ansible SSH keys for specific users
The previous method was installing the ansible SSH key in the root
home directory only.

Migrate that system and improve it to enable installing that key for
any user declared in a specific variable.

Add installation for root user for flash image and also add admin
user in other images to ease transitioning from root to admin for
driving test campaigns.

Issue: #5802
Change-Id: Ifc28fdbaf05d1c20cf2a024441d02cdc708f9379
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>