wic: fix /var ownership issues
The /var subdirectories are currently owned by the bitbake user, instead
of root. This is likely to be due to the `--rootfs-dir=` option used to
create the separate /var partition. The purpose of this partition is to
allow VMs to store data in a persistent partition separated from the
rootfs. Therefore:
* Delete the dedicated /var partition
* Add an empty /var/data partition for VMs to store their data

Issue: #5884
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Change-Id: I3440b9f535b61cefeba0f402e9e9ffcff5fc3a86