security: verify image configuration issues with PAM policy
Some images might declare undesired or unsafe Yocto features that
can impair or alter the PAM policy from operating properly.

Add additionnal checks so the image is skipped if incompatible features
are enabled.

Also add an image feature which allow the developer to explicitely
state that the image is relying on a unsafe PAM policy from SEAPATH's
perspective.

Issue: #5348
Change-Id: Ia16858f35faba29215e87f8138aba5fda3a7a96a
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>