security: fix installation of random-root-password
We should not rely on DISTRO_FEATURES to determine if systemd is
installed on the target as other init. systems can be prioritized
depending on other configuration items.

Verify that the init. system effectively installed is systemd while
also ensuring that the "sysinit.target" folder exists prior creating
the symlink.

Issue: #5337
Change-Id: I099dc9012f4a8324b6f2a13f6811f7b198004f08
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>