libvirt: set a static path for libvirtd configuration
By default, libvirtd tries to automatically find its configuration file,
depending if it is started as root or as user. If the service is started
by a user, libvirtd searches for a configuration in $XDG_CONFIG_HOME,
but this variable is not defined in the current system.
Thus, set libvirtd to always use the configuration file
located at /etc/libvirt/libvirtd.conf, which is the default root
configuration.

Issue: #5355
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Change-Id: I1b5b03eba74f11937fb34f2321c0c16542f28f4e