wic: do not override grub configuration with EFI disks
GRUB configuration is now signed and installed with "grub-efi" for
EFI images.

Enforce machines that can be used for guest and host images as WIC
is now relying on the bootloader installed, which is configured
depending on the machine used.

This is a step forward to have dedicated machines for EFI and Legacy
systems.

Issue: #5358
Change-Id: I3e6bbe29c1b2926cbb9f4249f625f679465ef455
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>