conf: increase meta-seapath layer priority
meta-efi-secure-boot has a priority of 10 which is higher than the
previous meta-seapath priority.

Increase it to "20" to ensure bbappend files and overrides coming from
meta-seapath are evaluated after those provided by meta-efi-secure-boot.

This will not affect the existing overrides as every other layer that
is used by meta-seapath has a strictly lower priority.

Issue: #5358
Change-Id: I5eba6be3c3c3d883c61c77bcd64c4e18e1f8e427
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>