security: enforce policy when changing password
"required" let the rest of the operation being done and add the failure
status to the final return value when all PAM related operations are
completed. Eventually, the user is notified that his password is not
valid, yet it has been changed in the database.

This is not desired as it is expected to abort if the user has supplied
an invalid password that does not match with the policy. Even worse, it
misguides the end-user by saying that the supplied password is rejected
while it actually isn't.

Enforce the abort by updating from "required" to "requisite" to stop
the password change attempt on an illegal password.

Also change the ordering of the inclusion to ensure that the PAM stack
is not reset and that a successful authentication will not skip the
cracklib module.

Issue: #5992
Change-Id: I7a4b0daf8dae34c1aed437f2fef8d689278fae1d
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>