security: add a class to generate traceability document
Based on traceability matrixes included by the layer, this class will
iterate over each matrix and look for information stored either in
runtime tests or in dedicated Yocto variable variable flags.

Implement two traceability parsers for "cukinia" and "yocto" that are
used to establish the traceability document artifact in asciidoc
format.

This artifact is then stored in DEPLOYDIR so it can be made available
for PDF generation at the end of the build.

This class is designed so multiple tracability matrixes can be declared
and parsed.

Issue: #5870
Change-Id: If303b494c754ab685ba6fd0d311686f573990133
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>