classes/boot-partition: add boot-partition.bbclass
Add boot-partition.bbclass with a task that permits the generation of
a .tar.gz with the contents of the bootloader and kernel.

This is required for the update system of the EFI images, where the
bootloader is flashed independently from the rootfs.

Note: for the moment, the rootfs will still contain a duplicated
unused /boot folder.

Issue: #5363
Signed-off-by: Albert Babí <albert.babi@savoirfairelinux.com>
Change-Id: I9bd664d0726975744683a5696c7b69cd62434d5b