classes/security/users: remove NOPASSWD
Remove 'NOPASSWD' from all created users and groups.
Add USERS_LIST_NOPASSWD and USERS_LIST_EXEC to let some users sudoers
file include NOPASSWD and EXEC tags, if needed. This is the case for
the ansible user, for example.
Add admin to these lists.
Add an exception to the cukinia test, so that the tests do not fail for
this specific user.
These privileges should only be attributed to the ansible user, once
the roles have been properly defined.

Issue: #5970
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Change-Id: I9a289277782292f5bcf7e7e3618fdb96ccd05b1b