libvirt: fix type error in default libvirt hook
The provided default libvirt hooks run external scripts as subprocesses
with their stdin connected to pipes, which expects bytes format by
default. The data about the started VM is then communicated as a
string, which creates a TypeError exception and ends the hook.

Thus, replace the default hook, with the only change being using pipes
in text mode to communicate with the subprocesses.

Issue: #5981
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Change-Id: I79c58f2deac93997b21963f46111afe7061c92d0