syslog-ng: change server bind address to localhost
Change syslog-ng server bind address to localhost.

Current configuration is limited to a specific implementation. It also
causes a bug where the syslog-ng server fails to start when the IP
address has not been set yet. Thus, this commit changes the default
syslog-ng server binding address to make sure it is always successfully
started. The configuration can then be modified with the correct binding
address.

Issue: #6184
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Change-Id: I6467e61317b8bf2dc0b5b87927a383d1518e271d