recipes-votp/system-config: always install a /sbin/init.sh
The init program has been changed to /sbin/init.sh in kernel parameters.
When using the seapath-ro distribution the file /sbin/init.sh will be
created. However, for the other distributions the the file will not be
created and the boot will fail.
TO fix the failure, this commit installs a /sbin/init.sh file which just
calls /sbin/init.

Issue: #6213
Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>
Change-Id: I92a6e74f703d330ddb378ee2b4a46a3de0ad20f1