security/user-config: add "emergadmin" account
This account is intended to be used for local debugging purposes.

It shall be the only account allowed for local logins. Rest of the
users will authenticate using SSH keys.

Add it to "privileged" group so it has all capabilities on the
system.

Also clarify variables documentation for USER_LIST_EXPIRED and
USER_LIST_SUDOERS.

Issue: #5359
Change-Id: I035d416619867ced72c7c97465d3d26e2769e9df
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>