security: remove no_login plugin from login policy
PAM plugin no_login helps in preventing any authentication other
than root when the file /etc/nologin is created.

But, the "emergadmin" account, which is the account to use in case
of catastrophic failure of the system shall never get denied from
logging on the system aside from invalid credentials.

Issue: #5892
Change-Id: Id9ba1bc0e58ea8fb565bb3293d5f687c712ca199
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>