switch_bootloader.sh: fix for some UEFI behavior
Some UEFI firmware tend to remove from the boot order all slots that are
explicitely disabled (if they are disabled, they have nothing to do in
the boot order).
switch_bootloader.sh assume both Seapath slots are present in the boot
order, but since the UEFI may have removed the disabled slot, this may
not be the case.
This fix makes switch_bootloader.sh working even if both slots are not
in the boot order.

Signed-off-by: insatomcat <florent.carli@rte-france.com>