kernel parameters: disable IPv6
IPv6 in SEAPATH has not been taken into account and tested. For security
reasons it is better to disable it.
The IPv6 disabling is done in kernel parameters to avoid givin the
possibility de re-enabled at runtime.
The IPv6 can easily be restored if necessary by deleting the kernel
parameter.

Signed-off-by: insatomcat <florent.carli@rte-france.com>