recipes-core/systemd: remove default DNS server
By default systemd-resolved, the DNS client use by SEAPATH use Goolge DNS servers is there no DNS server configured by systemd-networkd to always have a DNS server configured.

In SEAPATH the DNS servers configuration is done during the setup step is if during this step no DNS server was provided we condidered it is besause it is wanted and we do not want to failback to Google DNS servers.

Signed-off-by: insatomcat <florent.carli@rte-france.com>