ceph : bump version to 15.2.0 (dunfell default)
main branch was using a downgraded version of ceph (14.2.21). Since
there does not seem to be a reason not to use Dundell's default version
(ceph 15.2.0), this commit removes the ceph 14.2.21 recipe and makes the
default distro use dunfell default.
All specificities that had been coded in the ceph 14 recipes have been
ported to v15 with a ceph bbappend file.

Signed-off-by: insatomcat <florent.carli@rte-france.com>