Kernel config: add Macvtap driver and Vlan Tagging
MACVTAP: It is useful to be able to give direct network access to a guest on the
network card used for administration. The macvtap is the most efficient way for this.

VLAN TAGGING: Allows the kernel to be able to manage VLAN on interfaces
CONFIG_VLAN_8021Q addid to the kernel config.

Signed-off-by: insatomcat <florent.carli@rte-france.com>