init.sh: move kernel boot argument to machine (no-iommu)
Right now the init.sh script (used for overlay) is enabled only when the
readonly distro feature is enabled.
This changes allows the overlay to work whether readonly is enabled or
not.
A precedent commit already made that changes for host and monitor vm,
this commit is for the no-iommu machine

Signed-off-by: insatomcat <florent.carli@rte-france.com>