openssh: do generate ssh keys in a volatile directory
Use the same sshd_config for readonly as regular sshd_config.

There is no need to have sshd_config_readonly which differ from
sshd_config because, in SEAPATH we have an overlayfs in /etc and
modifications made in /etc are persistent, so it is possible to write
the SSH keys on it.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>