kernel parameters: remove hardcoded parameters for hugepages
The number of hugepages is no longer chosen in the seapath.conf.
This commit thus remove the "hugepages=X" kernel param.

Two others hugepages params have to be set for host only.
Remove a complicated way to set these params and just set it
in a conf file for hosts.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>