recipes-extended: add irqbalance
This commit package irq balance and add it to all hypervisor images.

Irqbalance is a daemon to help balance the cpu load generated by
interrupts across all of a systems cpus. Irqbalance identifies
the highest volume interrupt sources, and isolates each of them
to a single unique cpu, so that load is spread as much as possible
over an entire processor set, while minimizing cache miss rates.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>