distro: replace PREFERRED_VERSION by REQUIRED_VERSION
Previously, 'PREFERRED_VERSION' was used to specify the preferred
version of a recipe in the 'distro' module. With this if the version is
not found a warning will be raised and another version is used instead.
To avoid this, 'REQUIRED_VERSION' is now used instead of
'PREFERRED_VERSION'. This generates an error if the required version is
not found, instead of a warning. This change should help to prevent
unexpected behavior and make the code more robust.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>