classes/security/manifests-archiver: get kernel config from work-shared
Previously, the task of exporting kernel configuration was performed by
a task in the linux-mainline-rt recipe. However, this approach had
limitations - if the kernel was not rebuilt, for example, when the cache
was in use, the configuration was not exported.

To address this issue, we now retrieve the kernel configuration from the
work-shared directory and add a dependency to ensure that Yocto
recreates the kernel work-shared directory. This ensures that the kernel
configuration is always available and up to date.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>