classes/security/check-hardened-compilation: use checksec
The previous implementation of check-hardened-compilation do not check
all the hardened options and not support new options like
FORTITY_SOURCES=3.

This new implemention takes a lot of times to complete. To avoid loose
to much in developpement we add an option to skip it.
If SEAPATH_SECCOMPILE_MANIFEST_SKIP value is 1 the
check-hardened-compilation will be skiped.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>