classes/security/manifests-archiver: fix kernel config issue
The proper way to copy the kernel configuration inside the image
directory is to copy the file inside the kernel "DEPLOY_DIR" during the
"do_deploy" task. With this change the kernel configuration will be
stored inside the Yocto cache and will be automatically copy inside the
image directory before the manifest creation.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>