system-config: allow hawkbit server selection
Hawkbit server URL and port can be configured using the seapath.conf
file.
Use SEAPATH_HAWKBIT_SERVER_URL to define the hawkbit server URL.

For the record, the previous implementation used two variables:
HAWKBIT_SERVER_URL, which was in reality not the URL but only the domain
or IP part, and HAWKBIT_SERVER_PORT to define the port. I keep only one
variable that can be used to define the complete URL.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>