irqbalance: fix incorrect syntax usage
`IRQBALANCE_BANNED_CPULIST` is an environment variable from irqbalance
used to ban CPU core on IRQ threads. It can be used with a range syntax
such as `2-10` to exclude cores 2 to 10 on IRQ threads.

Currently, we used it using for example `4-N`, where `N` is the number
max of core available. This syntax is not supported by irqbalance,
causing to ignore the variable.

To fix it, this commit adds the call of a script in prelaunch task of
the irqbalance deamon, which retrieve the number max of CPU core
available and set it correctly in the irqbalance.env environnement
file.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>