irqbalance: fix set_irq_max_core.sh script
The set_irq_max_core.sh script is used to set the maximum core for
irqbalance to use. The script uses the nproc command to determine the
number of cores on the system. However, the nproc without any options
only returns the number of cores available to the current process. This
means that the script will not work correctly if the number of cores
available to the current process is less than the total number of cores
on the system for instance if isolcpus is used.

This patch changes the nproc command to use the --all option to return
the total number of cores on the system.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>