flash-script: extend persistent part to all space
Currently, size of the persistent partition is set by the WIC file. The
problem is if we want a bigger partition size, it will make final WIC
image bigger too, even if it contains no data.
To resolve this, this commit extends during flash the size of the
persistent partition to maximum freespace available on the disk.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>