cockpit: add recipe
Add Cockpit web UI support in Seapath Yocto. Cockpit is already
packaged in meta-openembedded on kirkstone branch, but is still using
the old 220 version.

As Seapath Debian use a latter version of Cockpit, this commit use the
last recipe available in master branch of meta-openembedded, which is
using Cockpit version 304.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>