linux-mainline-rt: enable DRM support as modules
Add the kernel configuration fragment drm.cfg, which adds the DRM
support for x86 machines. The DRM configurations are built as modules.

For an unknown reason, the DRM support also resolves an issue related to
libvirt. Without the DRM support, when libvirt starts or stops VMs, it
can freeze the host machine for a small but significant amount of time,
which can perturb real-time tasks.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>