openvswitch: add patch to remove hardcoded log file
The log file is hardcoded in the ovs-lib script. This patch removes the
hardcoded log file to allow changing the log file. This only affects
ovsdb-server.

Update the openvswitch.conf file to log only to syslog.
Remove /var/log/openvswitch from the ReadWritePaths in the ovsdb-server
service file as it is no longer used.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>