seapath-monitor-common.inc: install chrony on monitor
Timemaster is currently installed on both hosts and observers,
but chrony is only installed on hosts.
To make uniform which tools are used for PTP and NTP among machines in a
cluster, also install chrony and chronyc on the observer machine.
That way, the same timemaster/chrony configuration can be used for all
machines in a cluster.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>
Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>