snappy: re-enable compiling libsnappy with RTTI
Commit [1] disabled RTTI when compiling snappy. This means any program
using libsnappy won't be able to inherit classes defined by the library.
Especially, this broke snappy support in Ceph. [2]
As the snappy team refused to re-enable RTTI, most distributions are
patching snappy downstream to maintain support with applications
using it.

This adds a patch to re-enable building snappy with RTTI.

[1]: https://github.com/google/snappy/commit/c98344f6260d24d921e5e04006d4bedb528f404a
[2]: https://tracker.ceph.com/issues/53060

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>