python3-vm-manager: use github URL as URI
Initially vm-manager was fetched using repo and then use in Yocto with
local sources as it was not yet open-sourced on Github and to ease
development.

As vm-manager is now more stable, and to prepare for an official release
of Seapath, fetch vm-manager from its github repository instead of using
local sources.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>