python3-setup-ovs: use github URL as URI
Initially python3-setup-ovs was fetched using repo and then use in Yocto
with local sources as it was not yet open-sourced on Github and to ease
development.

As python3-setup-ovs is now more stable, and to prepare for an official
release of Seapath, fetch python3-setup-ovs from its github repository
instead of using local sources.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>