python3-vm-manager: fix RADOS permission issue
This patch fixes a permission denied issue when trying to connect to
the cluster with vm-manager.

If an "import libvirt" is done before performing the RADOS connection,
the connection will fail with a permission denied error.

As a workaround, we can create the RADOS connection when importing the
module and keep it alive until the end of the program.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>
Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>