cukinia/virtualization.conf: fix false positive
python3-pygments package, needed by netplan provides a file called
`spice.py`. This file is a lexer for spice protocol language, and is
not the spice protocol itself. So this file triggers a false positive.
To fix this, exclude the pygments subdirectory from the test to avoid
to find this file.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>