os-release: backport CPE_NAME field from scarthgap
Commit [1] added in os-release the CPE_NAME field which can then be
found in /etc/os-release. It can be used to detect that the distribution
was build with a yocto/openembedded environment.

As ansible seapath will use this field to detect if the distribution is
Yocto or not, backport it.

[1]: https://git.yoctoproject.org/poky/commit/?id=d674b5fe133ad83ea476eb9a2258683813e1be0f

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>