images: remove boot/ contents from the tarball
The wic image does not include the boot directory to allow mounting the
boot partition on the target. The tarball should reflect that.

Previously, after a swupdate, we could not remount the /boot partition
which made the swupdate_check.sh script not behave as expected. If
health checks failed, the grubenv file could not be updated.

We could consider moving this change to seapath-swupdate-common.inc,
but I preferred putting all the WIC related configuration in the same
file.

Signed-off-by: Enguerrand de Ribaucourt <enguerrand.de-ribaucourt@savoirfairelinux.com>