common_security_tests.d: refactor kernel configuration tests
This refactoring creates a common file containing the kernel
configurations to test on both x86 and arm. This file is used to define
the kernel configurations tests related to the architecture used.

This refactoring has no effect for x86 but removes the following tests
for arm, as they do not apply:
- PAGE_TABLE_ISOLATION:
- RETPOLINE:y
- LEGACY_VSYSCALL_NONE:y
- X86_VSYSCALL_EMULATION:n
- MODIFY_LDT_SYSCALL:n

SFL: #16303
Change-Id: I06ec39182988b39c002e9834e60f143ff07ddb85
Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>