hypervisor_tests.d: refactor kernel configuration tests
This refactoring creates a common file containing the kernel
configurations to test on both x86 and arm. This file is used to define
the kernel configurations tests related to the architecture used.

This refactoring has no effect for x86 but removes the following tests
for arm, as they do not apply:
- kernel_options["iommu"]
- kernel_options["hardware"] : X86_PKG_TEMP_THERMAL:y
- kernel_options["kvm"]

SFL: #16303
Change-Id: I4c88e4ed290b8758e56e7dfae246071a9e4e802c
Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>