u-boot-imx: add bootcount and bootlimit support
This enables bootcount and adds "bootlimit" uboot env variables.

The uboot bootcount env driver internally use "upgrade_available"
variable to increment bootcount.

If "upgrade_available" is 0, "bootcount" is always 1 in software env
if "upgrade_available" is 1 "bootcount" is incremented in the
environment and by default it is only saved on persistent storage when
it is greater than 1.

The userspace application must set the "upgrade_available" and
"bootcount" variable to 0, if a boot was successfully.

If bootcount is greater than bootlimit then uboot will run "altbootcmd"

SFL: #16301
Change-Id: Id437db097529516c575e61e573dac4401f5c97cd
Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>