bootcount.cfg: add environment configuration
These configurations ensure the u-boot env is stored on the eMMC (SD).

Currently, the u-boot env is stored on the external sd card. When the
time comes to flash the image to the internal mmc, the device declared
in fw_env.config should be updated, and the two following configurations
added:
- CONFIG_SYS_MMC_ENV_DEV=2
- CONFIG_SYS_MMC_ENV_PART=0

SFL: #16301
Change-Id: I60f37f9c31dd43b6c76871055679259e770abac7
Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>