security: Revert "security: disable pam_systemd.so with SSH connexion"
This reverts commit d70b579f1041fa470aec90fc76ffc62aa094efa0.
SSH sessions should pam_systemd.so plugin configured in common-session
so that the SSH session service and processes are attached to the user's
systemd slice.

Original problem with systemd was that if pam_systemd.so plugin is used,
user services will try to run on SSH session opening but will fail if
the user password is expired, even if an SSH key was used for login.
This was the case with the admin user, which previously set an unused
expired password by default. Previous commit fixed this behaviour.

Change-Id: I9b557ad79b6b8829d560a92083ca0869a5b3272b
Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>