security/users: add USERS_LIST_DISABLEMAXDAYS variable
Add USERS_LIST_DISABLEMAXDAYS to disable the password validity max days
value for added users in USERS_LIST.
Also add the admin user to USERS_LIST_DISABLEMAXDAYS.

When an SSH user session is opened, if sshd uses the pam_systemd.so
plugin, user services will be started. However, if the user's password
has expired the user services will fail to start even if an SSH key was
used to log in.
This is for example the case for the admin user. By default, SEAPATH
sets a maximum number of days of 90 for password validity.
For some reason users created with USERS_LIST are set with a
"last change date" from 2011, so the password is then immediately
expired for admin, even if password login is disabled with pam_access.

Change-Id: I3963cb067c586cbc26ce7dd53a2f7f99c3519ee4
Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>