aaeon-tools: fix undefined variable
This recipe is supposed to be used by the host container image only but
is also parsed when other images are used. As the variable
EXTRA_UBOOT_BOOTLOADER_FILE is added to SRC_URI and the latter is not
defined for machines not using u-boot, it generated a parsing error.

Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>