system-config: move ovs configuration to system-config.bb
OVS is currently installed on SEAPATH Debian standalone.
To harmonize both flavors python3-setup-ovs should also be installed
on SEAPATH Yocto standalone.

In order to do it in a future commit, move ovs and python3-setup-ovs
configuration from system-config-cluster to system-config to not
limit python3-setup-ovs configuration to cluster images.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>