python3-vm-manager: fix patch for standalone mode
vm-manager is supposed to detect if it can't load cluster related
packages and switch to standalone mode in that case.
However, 0001-vm-manager-fix-RADOS-permission-denied-issue.patch changes
were applied outside the try/except block providing this detection
mechanism. This means that on standalone hypervisors cluster would
execute and make vm-manager fail on initialization.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>