conf/distro/seapath-common.inc: raise an error if missing uid or gid
To avoid forgetting about adding static UID and GID for new groups and
users, set USERADD_ERROR_DYNAMIC to "error". This will raise a build
error when a user or group created by one of the built recipes doesn't
have an entry in passwd.uid and group.gid.

As this checks for users and groups of every built recipes, add static
mapping for additional user and group that weren't specified previously
even if they are not used in the final images.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>