users-config.inc: add user ansible
To harmonize configuration between Yocto and Debian SEAPATH, add the
user ansible which should be used with ansible playbooks.

For now, to preserve compatibility, the admin user is still created and
given the same ssh key as for the ansible user by keeping it in
USERS_SSH_ANSIBLE.
In the future, authentification for the admin and ansible users
should be uncorrelated.

Also update the cukinia tests.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>