security: access.conf: authorized connection for libvirt group
Since commit [1], the pam_access module was re-enabled for SSH
connections, including ssh sessions opened during livemigration.
However, in its current state, /etc/security/access.conf doesn't take
into account the livemigration user.

To re-enable livemigration, allow users of the libvirt group to log in.
Currently, the livemigration use is the only user in the libvirt group.

Also correct the user in group haclient in group.gid. Event if the users
associated in group.gid have no impact on the installation, this avoids
confusion as the root user is not part of haclient but livemigration is.

[1]: https://github.com/seapath/meta-seapath/commit/3e908292d1bb06d397ac7223f1f87f298613424d

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>