gnuplot: add x11 PACKAGECONFIG
This commit is s backport of [1], merged in master branch of
meta-openembedded.

Original commit message:
Currently, gnuplot recipe requires the x11 distro feature to be built.
However, X11 is not needed for running gnuplot as long as we don't use
a gnuplot X11 terminal.
This commit adds the x11 package config used to make the need of x11
distro feature optional to build gnuplot. `features_check` inherit is removed as
any of the functionalities of this class is now used.

[1]: https://github.com/openembedded/meta-openembedded/commit/7d46922407eb3d8009c592e61dccf4bb9eaf8df7

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>