conf/distro/include/group.gid: passwd.uid: avoid base-passwd conflicts
Recipe base-passwd from poky creates a set a groups and user
with a preconfigured passwd and group files instead of relying on Yocto
useradd and groupadd mechanisms. This is due to limitations explained
in one of the involved poky commit [1].

However, as base-passwd creates the sgx group with GID 48, this can
create a conflict when avahi is build because the avahi group was
assigned GID 48 in group.gid.

To avoid any conflicts between base-passwd and other packages:
* Change GIDs of groups with a GID below 100 to not interfere with the
  groups created by base-passwd.
* Add groups from base-passwd that were not yet in group.gid to avoid
  trying to use these GIDs in the future.

[1]: https://git.yoctoproject.org/poky/commit/?h=styhead&id=a10c01f9d16fc0bcdedde3cd51c6decd59e55263

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>