cockpit: use tarball instead of the npm class
This change switches the source of the cockpit-cluster-dashboard,
cockpit-cluster-vm-management, and cockpit-update recipes from using
the npm class to using a tarball from the respective GitHub releases.

This approach simplifies the build process by eliminating the need for
npm and nodejs during the build, and avoid having constantly updating
shrinkwrap files.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>